refactor: move frontend lib into src/lib
Fixes build/type errors in Vercel where frontend utilities outside src using import.meta.env were not consistently receiving Vite typings.

Keeps runtime frontend code inside src and improves project structure.
