+++ frontend/src/components/LoginButton.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { supabase } from '../../lib/supabaseClient';
+import { supabase } from '../lib/supabaseClient';
 import '../styles/components/auth-button.css';
 
 export const LoginButton = () => {
