+++ frontend/src/hooks/useAuthUser.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import type { User } from '@supabase/supabase-js';
-import { supabase } from '../../lib/supabaseClient';
+import { supabase } from '../lib/supabaseClient';
 
 export function useAuthUser() {
   const [user, setUser] = useState<User | null>(null);
