+++ frontend/src/lib/shouldSyncGuestData.ts
