+++ frontend/src/lib/supabaseClient.ts
@@ -1,3 +1,5 @@
+/// <reference types="vite/client" />
+
 import { createClient } from '@supabase/supabase-js';
 
 if (
